refactor NDArrayData to NDArray file
